+++ mod.ts
@@ -1,4 +1,3 @@
-export { nessieConfig, nessieConnection } from "./nessie.config.ts";
 export * from "./src/Column.ts";
 export * from "./src/Schema.ts";
 export * from "./src/Table.ts";
